Motor testing

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot_Parts/Servos.java
@@ -12,7 +12,7 @@ public class Servos {
         this.rightPlateServo = rightPlateServo;
     }
 
-    public void setPlateServoPos(int position) {
+    public void setPlateServoPos(double position) {
         rightPlateServo.setPosition(position);
         leftPlateServo.setPosition(position);
     }
